General Level Stuffs
Spawn Locations are a thing now. They're essentially just an object that allows you to size an area for the Enemy Spawner to use to spawn enemies (who would've thought?).

Speaking of Enemy Spawner, the Enemy Spawner spawns enemies (feels redundant to put into words). Enemy Spawner should be something that works implicitly and is created on scene transition (which we don't have yet because it's *only* 3 AM).

The other thing I did was make Room Info, which has nothing in it yet but the idea is for every scene to come with its own "metadata" which the level logic and enemy spawner can take from. So far, that includes the pool of enemies that can spawn, spawn locations, and references to the entrance and exit.

Other than that... just random tweaks like finally deleting that test object that has just been hanging out in the scene for the longest time. I think I also made another layer for colliders that shouldn't touch anything.

Now to lay down for a couple minutes and cry before animating the torso, coding the exit and entrance, importing all the tiles, and adding sounds. I hate everything.