New Tiles and Level Logic
RoomInfo class no longer exists because it was a dead end. All room objects will be placed there by hand (which is not that bad considering we'd have to still input the locations for them with my other idea).

Also, Unity's tools for JSON serialization are terrible and it makes me sad.