Attack Animations (WIP)
Attack animations are finished, they just don't activate yet

Also there's a glitch with the legs being flipped on their x-axis while the torso isn't and since the player is an uneven amount of pixels they become misaligned. This is because of a fix I made to the torso for the attack animations where it resets its x-axis when going up and down (because otherwise the sprite would be reaching in the wrong direction). The only way to fix this is to somehow make the legs also not do this. I'll probably do this tomorrow.