Lamp Animations
I'm going insane

The pixel perfect camera moves everything to the closest pixel, which means if the lamp was even slightly off relative to the player, it would move out of sync with the player. So for all 16 attack animations I had to find the exact relative distance from the player. I haven't even found out a way to keep the lantern behind the player during certain attack animations since you can't change layers in the animator.

This took me 3 hours... and it didn't even accomplish much